refactor(model)!: move datetime to util
Also remove the `datetime::error` module to be consistent with other
error paths.

`model::datetime::Timestamp` becomes `model::util::Timestamp`, with its
errors and display under `model::util::datetime`
