+++ cache/in-memory/src/model/guild.rs
@@ -2,7 +2,6 @@ use std::slice::Iter;
 
 use serde::Serialize;
 use twilight_model::{
-    datetime::Timestamp,
     guild::{
         DefaultMessageNotificationLevel, ExplicitContentFilter, MfaLevel, NSFWLevel, Permissions,
         PremiumTier, SystemChannelFlags, VerificationLevel,
@@ -11,7 +10,7 @@ use twilight_model::{
         marker::{ApplicationMarker, ChannelMarker, GuildMarker, UserMarker},
         Id,
     },
-    util::image_hash::ImageHash,
+    util::{ImageHash, Timestamp},
 };
 
 /// Represents a cached [`Guild`].
