+++ cache/in-memory/src/permission.rs
@@ -530,7 +530,6 @@ mod tests {
             permission_overwrite::{PermissionOverwrite, PermissionOverwriteType},
             Channel, ChannelType,
         },
-        datetime::Timestamp,
         gateway::payload::incoming::{
             ChannelCreate, GuildCreate, MemberAdd, MemberUpdate, RoleCreate,
         },
@@ -542,6 +541,7 @@ mod tests {
             marker::{ChannelMarker, GuildMarker, RoleMarker, UserMarker},
             Id,
         },
+        util::Timestamp,
     };
 
     assert_fields!(ChannelErrorType::ChannelUnavailable: channel_id);
