+++ http/src/request/scheduled_event/create_guild_scheduled_event/stage_instance.rs
@@ -5,9 +5,9 @@ use crate::{
     response::ResponseFuture,
 };
 use twilight_model::{
-    datetime::Timestamp,
     id::{marker::ChannelMarker, Id},
     scheduled_event::{EntityType, GuildScheduledEvent},
+    util::Timestamp,
 };
 use twilight_validate::request::{
     audit_reason as validate_audit_reason,
