+++ model/src/application/interaction/mod.rs
@@ -410,11 +410,11 @@ mod test {
                 Interaction, InteractionType,
             },
         },
-        datetime::{Timestamp, TimestampParseError},
         guild::{PartialMember, Permissions},
         id::Id,
         test::image_hash,
         user::User,
+        util::datetime::{Timestamp, TimestampParseError},
     };
     use serde_test::Token;
     use std::{collections::HashMap, str::FromStr};
