+++ model/src/channel/reaction.rs
@@ -180,11 +180,11 @@ impl<'de> Deserialize<'de> for Reaction {
 mod tests {
     use super::super::{Reaction, ReactionType};
     use crate::{
-        datetime::{Timestamp, TimestampParseError},
         guild::Member,
         id::Id,
         test::image_hash,
         user::User,
+        util::datetime::{Timestamp, TimestampParseError},
     };
     use serde_test::Token;
     use std::str::FromStr;
