+++ model/src/lib.rs
@@ -71,7 +71,6 @@
 
 pub mod application;
 pub mod channel;
-pub mod datetime;
 pub mod gateway;
 pub mod guild;
 pub mod http;
