+++ util/src/builder/interaction_response_data.rs
@@ -178,7 +178,7 @@ mod tests {
     use std::fmt::Debug;
     use twilight_model::{
         application::component::{button::ButtonStyle, Button},
-        datetime::Timestamp,
+        util::Timestamp,
     };
 
     assert_impl_all!(
